chore: re-arrange types
